pyproject.toml
[project]
name = "sym-backend"
version = "0.1.0"
description = "Save your moments"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.2",
    "django-environ==0.12.0",
    "django-split-settings==1.3.2",
    "djangorestframework==3.16.0",
]
